refactor getCreds
